Correcciones

+++ src/main/java/hn/unah/bases/proyecto_playstore/IngresarDatosDestino.java
@@ -7,18 +7,21 @@ import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.ArrayList;
 
-import javax.sql.rowset.CachedRowSet;
 
-import hn.unah.bases.proyecto_playstore.DTOs.CampoDTO;
 
 public class IngresarDatosDestino {
 
     public int ingresarRegistros(Connection connOrg, Connection connDes,
-            ArrayList<String> camposDestino, ArrayList<String> camposOrigen, String tableOrigen, String tableDestino) {
+            ArrayList<String> camposDestino, ArrayList<String> camposOrigen, String tableOrigen, String tableDestino, boolean fromTable) {
+
 
         int cantInsert = 0;
 
         try {
+
+            //eliminamos los registros de la tabla para volverlos a ingresar
+            PreparedStatement stmtDelete = connDes.prepareStatement("delete "+ tableDestino);
+            stmtDelete.executeUpdate();
 
             // Construir la parte de la consulta SQL para seleccionar los campos de origen
             StringBuilder camposConsulta = new StringBuilder();
@@ -68,9 +71,12 @@ public class IngresarDatosDestino {
                 cantInsert++;
             }
 
-            //DEJAMOS CAER LA TABLA
-			PreparedStatement stmtDrop=connOrg.prepareStatement("drop table temp_table");
-			stmtDrop.executeUpdate();
+            if(!fromTable){
+                 //DEJAMOS CAER LA TABLA
+                PreparedStatement stmtDrop=connOrg.prepareStatement("drop table temp_table");
+                stmtDrop.executeUpdate();
+            }
+           
 
           
             // Cerrar la conexión cuando hayas terminado
@@ -86,44 +92,6 @@ public class IngresarDatosDestino {
         return cantInsert;
     }
 
-    public int ingresarRegistros2(Connection connDes,
-            CachedRowSet registrosConvertidos, String[] orden, String tableDestino) {
-
-        int cantInsert = 0;
-
-        try {
-            registrosConvertidos.first();
-
-            StringBuilder consultaInsert = new StringBuilder("INSERT INTO " + tableDestino + " VALUES (");
-            for (int i = 1; i <= orden.length; i++) {
-
-                consultaInsert.append("?,");
-            }
-            consultaInsert.deleteCharAt(consultaInsert.length() - 1); // Eliminar la última coma
-            consultaInsert.append(")");
-            PreparedStatement stmtInsert = connDes.prepareStatement(consultaInsert.toString());
-
-            // Iterar sobre los resultados utilizando el cursor
-            do {
-                // Insertar los valores en la tabla de destino
-                for (int i = 1; i <= orden.length; i++) {
-                    int index = Integer.parseInt(orden[i - 1]);
-
-                    stmtInsert.setObject(i, registrosConvertidos.getObject(index));
-
-                }
-
-                // Ejecutar la consulta de inserción
-                stmtInsert.executeUpdate();
-                cantInsert++;
-            } while (registrosConvertidos.next());
-
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-
-        return cantInsert;
-    }
 
     public void alterMaxLength(Connection connDes, String table, String campo, String cambio) throws SQLException {
         String consulta = "ALTER TABLE " + table + " MODIFY " + campo + " " + cambio;
